improvment, anti DoS for ask sync processing.
